rename components to systems
